refactor: web 핸들러 패키지 구체화
