+++ src/main/java/roomescape/controller/common/AuthController.java
@@ -9,11 +9,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.config.LoginMember;
-import roomescape.dto.response.MemberNameResponse;
 import roomescape.dto.request.LoginRequest;
+import roomescape.dto.response.MemberNameResponse;
 import roomescape.entity.AccessToken;
 import roomescape.service.MemberService;
+import roomescape.web.LoginMember;
 
 @RestController
 @RequestMapping("/auth")
