+++ src/main/java/roomescape/controller/user/ReservationController.java
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.config.LoginMember;
 import roomescape.dto.request.ReservationPostRequestByUser;
 import roomescape.dto.response.ReservationPostResponse;
 import roomescape.service.ReservationCommandService;
+import roomescape.web.LoginMember;
 
 @RestController
 @RequestMapping("/reservations")
