+++ src/main/java/roomescape/web/support/CheckAdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.web.support;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
